# cmd/api/docs/swagger.yaml
definitions:
...
  types.ClassResponse:
    properties:
      description:
        type: string
      google_course_id:
        type: string
      google_course_link:
        type: string
      google_synced_at:
        type: string
      id:
        type: integer
      owner_id:
        type: integer
      status:
        type: integer
      topic:
        type: string
    type: object
...
  types.CreateClassRequest:
    properties:
      description:
        type: string
      google_course_id:
        type: string
      google_course_link:
        type: string
      status:
        description: public or private
        type: integer
      topic:
        type: string
    required:
    - description
    - topic
    type: object
...
  types.MemberResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      picture_path:
        type: string
    type: object
...
  types.UpdateClassRequest:
    properties:
      description:
        type: string
      google_course_id:
        type: string
      google_course_link:
        type: string
      status:
        description: public or private
        type: integer
      topic:
        type: string
    type: object
...
info:
  contact: {}
  description: This is the API documentation for the Assembly Visual project.
  termsOfService: http://swagger.io/terms/
  title: Assembly Visual API documentation
  version: "1.0"
paths:
...
  /classroom:
    get:
      consumes:
      - application/json
      description: Retrieve all classrooms
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/types.ClassResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      tags:
      - classrooms
    post:
      consumes:
      - application/json
      description: Create a new class
      parameters:
      - description: Class info
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/types.CreateClassRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      tags:
      - classrooms
  /classroom/{class_id}:
    delete:
      consumes:
      - application/json
      description: Delete a classroom
      parameters:
      - description: Class ID
        in: path
        name: class_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      tags:
      - classrooms
    get:
      consumes:
      - application/json
      description: Retrieve a class by ID
      parameters:
      - description: Class ID
        in: path
        name: class_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ClassResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      tags:
      - classrooms
    put:
      consumes:
      - application/json
      description: Update an existing class
      parameters:
      - description: Class ID
        in: path
        name: class_id
        required: true
        type: integer
      - description: Updated class info
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/types.UpdateClassRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      tags:
      - classrooms
...
  /classroom/{class_id}/join:
    post:
      consumes:
      - application/json
      description: Join a classroom
      parameters:
      - description: Class ID
        in: path
        name: class_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      tags:
      - classrooms
  /classroom/{class_id}/members:
    get:
      consumes:
      - application/json
      description: Get all members of a class by class ID
      parameters:
      - description: Class ID
        in: path
        name: class_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/types.MemberResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      tags:
      - classrooms
  /classroom/public:
    get:
      consumes:
      - application/json
      description: Retrieve all public classrooms
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/types.ClassResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      tags:
      - classrooms
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"